[refactor] project structure following fastapi template
