+++ backend/app/main.py
@@ -1,5 +1,5 @@
-from domain.user import user_router
-from domain.chat import chat_router
+from app.api.routes import user_router
+from app.api.routes import chat_router
 
 from fastapi import FastAPI
 from starlette.middleware.cors import CORSMiddleware
@@ -7,8 +7,8 @@ from starlette.responses import FileResponse
 from dotenv import load_dotenv
 import os
 load_dotenv()
-
 DEV_FRONTEND_URL = os.getenv("DEV_FRONTEND_URL")
+
 origins = [
     DEV_FRONTEND_URL
 ]
