+++ backend/scripts/prestart.sh
@@ -4,4 +4,4 @@ alembic upgrade head
 
 ollama serve &
 
-uvicorn main:app --host 0.0.0.0 --port 8000
+uvicorn app.main:app --host 0.0.0.0 --port 8000
